New milestone in styling for the index.php page
Layout is streamlined.

+++ src/js/karaoke_table.js
@@ -14,6 +14,7 @@ $(document).ready(function () {
     * content of the 'custom_filter_section'
     */
         search_icon = '<img src="images/search_icon.png" width="32px" height="32px" alt="search icon" id="search-icon" />',
+        search_reset_button = '<button id="search_reset" class="filter-reset-button">X</button>',
     /*
     * END
     * content of the 'custom_filter_section'
@@ -199,7 +200,7 @@ $(document).ready(function () {
                 data.registered = registered;
 			}
 		},
-		"dom": "<\"col-xs-12\"f<\"#search_reset\">><\"col-xs-12 navbar navbar-default\" <\"custom-filter-container container-fluid\">><\"col-xs-12\"i><\"col-sm-8 col-xs-12\"l><\"col-sm-4 col-xs-12\"p><\"col-xs-12\"t><\"col-sm-4 col-xs-12\"i><\"col-sm-8 col-xs-12\"p><\"col-xs-12\"l>r",
+		"dom": "f<\"col-xs-12 navbar navbar-default\" <\"custom-filter-container container-fluid\">><\"col-xs-12\"i><\"col-sm-8 col-xs-12\"l><\"col-sm-4 col-xs-12\"p><t><\"col-sm-4 col-xs-12\"i><\"col-sm-8 col-xs-12\"p><\"col-xs-12\"l>r",
 		"responsive" : true,
 		"columnDefs": [
             {"orderable": false, "targets": [3, 10]},
@@ -216,9 +217,6 @@ $(document).ready(function () {
 		"sPaginationType": "listbox",
         "drawCallback": prepareVideosTable
     }).fnFilterOnReturn();
-    
-    //assign a title-attribute to the #search-reset element
-    $('#search_reset').attr('title', 'Reset Search').html('X');
     
     //assign HTML content to the .custom-filter-container class
     $('div.custom-filter-container').html(custom_filter_section);
@@ -380,13 +378,13 @@ $(document).ready(function () {
         //"Search" text of the
         //input search field
         $('#videos_datatable_filter.dataTables_filter label')
-        .contents()
-        .filter(function () {
-            //find the only node item
-            //that is of type 'text'
-            return this.nodeType === 3; 
-        })
-        .remove();
+            .contents()
+            .filter(function () {
+                //find the only node item
+                //that is of type 'text'
+                return this.nodeType === 3; 
+            })
+            .remove();
         
         //add a custom PNG
         //search icon to
@@ -394,6 +392,14 @@ $(document).ready(function () {
         //the input search field
         if ($('#search-icon').length === 0) {
             $('#videos_datatable_filter.dataTables_filter label').prepend(search_icon);    
+        }
+        
+        //add a custom 
+        //'search reset' button
+        //right after
+        //the input search field
+        if ($('#search_reset').length === 0) {
+            $('#videos_datatable_filter.dataTables_filter label').append(search_reset_button);    
         }
         
         //grab the values/options
@@ -457,10 +463,15 @@ $(document).ready(function () {
                 thisRow.attr('data-index_position', video_index - 1);
                 var rowData = videos_datatable.row(thisRow).data();
                 addOfflineNotif('videos_datatable', rowData, thisRow);
-                
-                /*('<p>' + indexdata + '</p><a href="update_video.php?id=' + video_index + '"><p>Edit Video Details</p></a><a href="delete_video.php?video_id=' + videoiddata + '&song_title=' + songtitle_join + '" onclick="return confirm(\'Are You Sure?\')"><p>Delete Video</p></a>');*/
             });    
         } 
+        
+        //make the search reset button
+        //functional
+        $('#search_reset').on('click', function () {
+            $('.form-control.input-sm').val('');
+            videos_datatable.search('').draw();
+        });
     }
     /*
     * functions to be invoked
@@ -479,26 +490,28 @@ $(document).ready(function () {
     if (registered === 'green') {
         //on every videos table render
         videos_datatable.on('draw.dt', function () {
-            $('#videos_datatable tbody tr').each(function () {
-                //assign helper vars
-                var thisRow = $(this),
-                    rowData = videos_datatable.row(thisRow).data(),
-                    song_db_id = rowData[1],
-                    songtitle = rowData[0],
-                    songtitlesplit = rowData[0].split(' '),
-                    songtitlejoin = songtitlesplit.join('+'),
-                    //get the datatable row Node object
-                    firstNode = videos_datatable.cell(this, 0).node();
-                
-                //assemble the links
-                //to be added to each table row
-                var crudLinks = '<p>' + songtitle + '</p>';
-                    crudLinks += '<a href="update_video.php?id=' + song_db_id + '">Edit Video Details</a>';
-                    crudLinks += '<a href="delete_video.php?video_id=' + song_db_id + '&song_title=' + songtitlejoin + '" onclick="return confirm(\'Are You Sure?\')"><p>Delete Video</p></a>';
-                
-                //add the links
-                $(firstNode).html(crudLinks);
-            });
+            if ($('#videos_datatable tbody tr td.dataTables_empty').length < 1) {
+                $('#videos_datatable tbody tr').each(function () {
+                    //assign helper vars
+                    var thisRow = $(this),
+                        rowData = videos_datatable.row(thisRow).data(),
+                        song_db_id = rowData[1],
+                        songtitle = rowData[0],
+                        songtitlesplit = rowData[0].split(' '),
+                        songtitlejoin = songtitlesplit.join('+'),
+                        //get the datatable row Node object
+                        firstNode = videos_datatable.cell(this, 0).node();
+
+                    //assemble the links
+                    //to be added to each table row
+                    var crudLinks = '<p>' + songtitle + '</p>';
+                        crudLinks += '<a href="update_video.php?id=' + song_db_id + '">Edit Video Details</a>';
+                        crudLinks += '<a href="delete_video.php?video_id=' + song_db_id + '&song_title=' + songtitlejoin + '" onclick="return confirm(\'Are You Sure?\')"><p>Delete Video</p></a>';
+
+                    //add the links
+                    $(firstNode).html(crudLinks);
+                });    
+            }
         });
     }
     /*
@@ -537,20 +550,6 @@ $(document).ready(function () {
             $('#info-buttons ul li#showinfo').show();
         }
     });
-    
-    //make the search reset button
-    //functional
-    $('#search_reset').on('click', function () {
-        $('.form-control.input-sm').val('');
-        videos_datatable.search('').draw();
-    });
-	
-    //clicking the 'Reset' button
-    //beside the search input
-    //will reset the table
-	$('#search-reset').click(function () {
-		videos_datatable.search('').draw();
-	});
 	
 	//Custom filters
 	$('#decade-filter').change(function () {
